Prototype:  `eval_in_place` + Benchmarks
Inspired by the "Avoid cloning where possible":
https://github.com/gifnksm/polynomial-rs/issues/27

First, let's add a benchmark so we can see if avoiding cloning helps
with performance.

Second, add an `eval_in_place` method that avoids the allocation, and
see if that actually improves performance on types that store data on
the heap.

Conclusion:  it doesn't seem like this matters, at least for the
`BigRational` data types.